Consolidate more components
