+++ airflow-core/src/airflow/ui/src/components/ui/ActionButton.tsx
@@ -16,10 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { type ButtonProps, IconButton } from "@chakra-ui/react";
+import { Button, type ButtonProps, IconButton } from "@chakra-ui/react";
 import type { FC, ReactElement } from "react";
 
-import { Button, Tooltip } from "src/components/ui";
+import { Tooltip } from "src/components/ui";
 
 type Props = {
   readonly actionName: string;
