+++ airflow-core/src/airflow/ui/src/components/ui/Select/Trigger.tsx
@@ -16,10 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { Select as ChakraSelect } from "@chakra-ui/react";
+import { CloseButton, Select as ChakraSelect } from "@chakra-ui/react";
 import { forwardRef } from "react";
-
-import { CloseButton } from "../CloseButton";
 
 type Props = {
   readonly clearable?: boolean;
