+++ airflow-core/src/airflow/ui/src/components/ui/index.ts
@@ -21,13 +21,11 @@ export * from "./Dialog";
 export * from "./Pagination";
 export * from "./Select";
 export * from "./Alert";
-export * from "./CloseButton";
 export * from "./Switch";
 export * from "./Tooltip";
 export * from "./ProgressBar";
 export * from "./Menu";
 export * from "./Accordion";
-export * from "./Button";
 export * from "./Toaster";
 export * from "./Breadcrumb";
 export * from "./Clipboard";
