+++ airflow-core/src/airflow/ui/src/pages/Connections/DeleteConnectionsButton.tsx
@@ -16,12 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { Code, Flex, Heading, Text, VStack, useDisclosure } from "@chakra-ui/react";
+import { Button, Code, Flex, Heading, Text, useDisclosure, VStack } from "@chakra-ui/react";
 import { useTranslation } from "react-i18next";
 import { FiTrash2 } from "react-icons/fi";
 
 import { ErrorAlert } from "src/components/ErrorAlert";
-import { Button, Dialog } from "src/components/ui";
+import { Dialog } from "src/components/ui";
 import { useBulkDeleteConnections } from "src/queries/useBulkDeleteConnections";
 
 type Props = {
